refactor: implement architecture improvements
- Move JoinView.tsx and ResultsPage.tsx to components/views/
- Split gameHandler.js into focused sub-handlers:
  - gameLifecycleHandler.js (create, start, end, reset)
  - playerJoinHandler.js (join, leave, reconnection)
  - roomManagementHandler.js (close room, active rooms, grid shuffling)
- Update imports to use @/ path alias consistently
- Add comprehensive integration tests for split handlers

These changes improve code organization and maintainability
while maintaining full backward compatibility.
