+++ fe-next/components/views/index.ts
@@ -1,0 +1,7 @@
+/**
+ * Views Components
+ * Top-level view components for the application
+ */
+
+export { default as JoinView } from './JoinView';
+export { default as ResultsPage } from './ResultsPage';
